tidy up error functions, separate comment endpoints
