+++ error-messages/errors.js
@@ -1,0 +1,18 @@
+exports.tooManyProps = () => {
+  return Promise.reject({
+    status: 400,
+    msg: "only updates to inc_votes are available",
+  });
+};
+
+exports.idNotFound = () => {
+  return Promise.reject({ status: 404, msg: "review_id not found" });
+};
+
+exports.idNotNumber = () => {
+  return Promise.reject({ status: 400, msg: "review_id must be a number" });
+};
+
+exports.incVotesNotNumber = () => {
+  return Promise.reject({ status: 400, msg: "inc_votes must be a number" });
+};
